requirements.txt
allure-pytest==2.13.2
allure-python-commons==2.13.2
attrs==23.2.0
certifi==2023.11.17
charset-normalizer==3.3.2
colorama==0.4.6
et-xmlfile==1.1.0
exceptiongroup==1.2.0
idna==3.6
iniconfig==2.0.0
numpy==1.26.3
openpyxl==3.1.2
packaging==23.2
pandas==2.1.4
pluggy==1.3.0
pytest==7.4.4
python-dateutil==2.8.2
pytz==2023.3.post1
requests==2.31.0
simplejson==3.19.2
six==1.16.0
tomli==2.0.1
tzdata==2023.4
urllib3==2.1.0
